Agente Educador/pyproject.toml
[tool.poetry]
name = "gemini-educador"
version = "0.1.0"
description = "Plataforma educativa interactiva para niños"
authors = ["Tu Nombre <tu@email.com>"]
packages = [{ include = "app" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.2"
pydantic = "^1.10.12"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pymongo = "^4.13.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
email-validator = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\\.pyi?$'
